+++ src/Shader/Shader.hpp
@@ -552,7 +552,7 @@ namespace sw
 		void append(Instruction *instruction);
 		void declareSampler(int i);
 
-		const Instruction *getInstruction(unsigned int i) const;
+		const Instruction *getInstruction(size_t i) const;
 		int size(unsigned long opcode) const;
 		static int size(unsigned long opcode, unsigned short version);
